requirements.txt
scikit-learn>=1.0.0
numpy>=1.20.0
pdfminer.six>=20220524
python-docx>=0.8.11 